// src/alerts/coinbase/rest/types/common-types.ts
export enum MarginType {
  CROSS = 'CROSS',
  ISOLATED = 'ISOLATED',
}

export enum OrderPlacementSource {
  UNKNOWN = 'UNKNOWN_PLACEMENT_SOURCE',
  RETAIL_SIMPLE = 'RETAIL_SIMPLE',
  RETAIL_ADVANCED = 'RETAIL_ADVANCED',
}

export enum SortBy {
  UNKNOWN = 'UNKNOWN_SORT_BY',
  LIMIT_PRICE = 'LIMIT_PRICE',
  LAST_FILL_TIME = 'LAST_FILL_TIME',
}

export enum OrderSide {
  BUY = 'BUY',
  SELL = 'SELL',
}
// ...
export type MarketMarketIoc = { quote_size: string } | { base_size: string };
// ...
export type FCMBalanceSummary = {
  futures_buying_power?: Record<string, any>;
  total_usd_balance?: Record<string, any>;
  cbi_usd_balance?: Record<string, any>;
  cfm_usd_balance?: Record<string, any>;
  total_open_orders_hold_amount?: Record<string, any>;
  unrealized_pnl?: Record<string, any>;
  daily_realized_pnl?: Record<string, any>;
  initial_margin?: Record<string, any>;
  available_margin?: Record<string, any>;
  liquidation_threshold?: Record<string, any>;
  liquidation_buffer_amount?: Record<string, any>;
  liquidation_buffer_percentage?: string;
  intraday_margin_window_measure?: Record<string, any>;
  overnight_margin_window_measure?: Record<string, any>;
};

export type FCMPosition = {
  product_id?: string;
  expiration_time?: Record<string, any>;
  side?: Record<string, any>;
  number_of_contracts?: string;
  current_price?: string;
  avg_entry_price?: string;
  unrealized_pnl?: string;
  daily_realized_pnl?: string;
};
// ...
export interface TradingState {
  lastTradeTime: number;
  dailyLoss: number;
  openPositions: number;
  currentPosition?: FCMPosition;
  lastAlert: any | null;
  balanceSummary?: FCMBalanceSummary;
}

export interface TradingSignal {
  action: OrderSide | null;
  isValid: boolean;
}